+++ packages/shared/src/apis/HTTPError.ts
