+++ packages/shared/src/constants/api.ts
@@ -1,0 +1,7 @@
+export const HTTP_METHOD = {
+  get: 'GET',
+  post: 'POST',
+  patch: 'PATCH',
+  delete: 'DELETE',
+  put: 'PUT',
+} as const;
